feat/SW-38/refactor(Like): PostgreSQL 마이그레이션에 따른 Like 도메인 수정
* Note: 추후 확장 기능이기 때문에 일부만 마이그레이션

+++ src/main/java/com/web/SearchWeb/likes/dao/LikesDao.java
@@ -1,16 +1,15 @@
 package com.web.SearchWeb.likes.dao;
 
-
 public interface LikesDao {
     // 게시글 좋아요 상태 확인
-    Boolean isLikedByMember( int boardId, int memberId);
+    Boolean isLikedByMember(Long boardId, Long memberId);
 
     // 게시글 좋아요 추가
-    int likeBoard(int boardId, int memberId);
+    int likeBoard(Long boardId, Long memberId);
 
     // 게시글 좋아요 취소
-    int unlikeBoard(int boardId, int memberId);
+    int unlikeBoard(Long boardId, Long memberId);
 
     // 게시글 좋아요 수 조회
-    int countLikes(int boardId);
+    int countLikes(Long boardId);
 }
